+++ activerecord/test/cases/adapters/abstract_mysql_adapter/optimizer_hints_test.rb
@@ -11,7 +11,7 @@ class OptimizerHintsTest < ActiveRecord::AbstractMysqlTestCase
       assert_queries_match(%r{\ASELECT /\*\+ NO_RANGE_OPTIMIZATION\(posts index_posts_on_author_id\) \*/}) do
         posts = Post.optimizer_hints("NO_RANGE_OPTIMIZATION(posts index_posts_on_author_id)")
         posts = posts.select(:id).where(author_id: [0, 1])
-        assert_includes posts.explain, "| index | index_posts_on_author_id | index_posts_on_author_id |"
+        assert_includes posts.explain.inspect, "| index | index_posts_on_author_id | index_posts_on_author_id |"
       end
     end
 
@@ -27,7 +27,7 @@ class OptimizerHintsTest < ActiveRecord::AbstractMysqlTestCase
       assert_queries_match(%r{\ASELECT /\*\+ NO_RANGE_OPTIMIZATION\(posts index_posts_on_author_id\) \*/}) do
         posts = Post.optimizer_hints("/*+ NO_RANGE_OPTIMIZATION(posts index_posts_on_author_id) */")
         posts = posts.select(:id).where(author_id: [0, 1])
-        assert_includes posts.explain, "| index | index_posts_on_author_id | index_posts_on_author_id |"
+        assert_includes posts.explain.inspect, "| index | index_posts_on_author_id | index_posts_on_author_id |"
       end
 
       assert_queries_match(%r{\ASELECT /\*\+ \*\* // `posts`\.\*, // \*\* \*/}) do